Imported Upstream version 2.0.28+dfsg

+++ aqt/main.py
@@ -16,7 +16,7 @@ import aqt.progress
 import aqt.webview
 import aqt.toolbar
 import aqt.stats
-from aqt.utils import  restoreGeom, showInfo, showWarning,\
+from aqt.utils import saveGeom, restoreGeom, showInfo, showWarning, \
     restoreState, getOnlyText, askUser, applyStyles, showText, tooltip, \
     openHelp, openLink, checkInvalidFilename
 import anki.db
@@ -271,14 +271,13 @@ To import into a password protected profile, please open the profile before atte
             self.col = Collection(cpath, log=True)
         except anki.db.Error:
             # warn user
-            showWarning("""\
-Your collection is corrupt. Please see the manual for \
-how to restore from a backup.""")
-            # move it out of the way so the profile can be used again
-            newpath = cpath+str(intTime())
-            os.rename(cpath, newpath)
-            # then close
-            sys.exit(1)
+            showWarning(_("""\
+Your collection is corrupt. Please create a new profile, then \
+see the manual for how to restore from an automatic backup.
+
+Debug info:
+""")+traceback.format_exc())
+            self.unloadProfile()
         except Exception, e:
             # the custom exception handler won't catch this if we immediately
             # unload, so we have to manually handle it
@@ -304,12 +303,17 @@ how to restore from a backup.""")
         if self.col:
             if not self.closeAllCollectionWindows():
                 return
-            self.maybeOptimize()
             self.progress.start(immediate=True)
-            if os.getenv("ANKIDEV", 0):
-                corrupt = False
-            else:
-                corrupt = self.col.db.scalar("pragma integrity_check") != "ok"
+            corrupt = False
+            try:
+                self.maybeOptimize()
+            except:
+                corrupt = True
+            if not corrupt:
+                if os.getenv("ANKIDEV", 0):
+                    corrupt = False
+                else:
+                    corrupt = self.col.db.scalar("pragma integrity_check") != "ok"
             if corrupt:
                 showWarning(_("Your collection file appears to be corrupt. \
 This can happen when the file is copied or moved while Anki is open, or \
@@ -764,9 +768,9 @@ title="%s">%s</button>''' % (
         import aqt.importing
         aqt.importing.onImport(self)
 
-    def onExport(self):
+    def onExport(self, did=None):
         import aqt.exporting
-        aqt.exporting.ExportDialog(self)
+        aqt.exporting.ExportDialog(self, did=did)
 
     # Cramming
     ##########################################################################
@@ -971,7 +975,9 @@ will be lost. Continue?"""))
         diag.connect(box, SIGNAL("rejected()"), diag, SLOT("reject()"))
         diag.setMinimumHeight(400)
         diag.setMinimumWidth(500)
+        restoreGeom(diag, "checkmediadb")
         diag.exec_()
+        saveGeom(diag, "checkmediadb")
 
     def deleteUnused(self, unused, diag):
         if not askUser(
@@ -980,7 +986,8 @@ will be lost. Continue?"""))
         mdir = self.col.media.dir()
         for f in unused:
             path = os.path.join(mdir, f)
-            send2trash(path)
+            if os.path.exists(path):
+                send2trash(path)
         tooltip(_("Deleted."))
         diag.close()
 
@@ -1003,10 +1010,12 @@ will be lost. Continue?"""))
         self.progress.finish()
         part1 = ngettext("%d card", "%d cards", len(cids)) % len(cids)
         part1 = _("%s to delete:") % part1
-        diag, box = showText(part1 + "\n\n" + report, run=False)
+        diag, box = showText(part1 + "\n\n" + report, run=False,
+                geomKey="emptyCards")
         box.addButton(_("Delete Cards"), QDialogButtonBox.AcceptRole)
         box.button(QDialogButtonBox.Close).setDefault(True)
         def onDelete():
+            saveGeom(diag, "emptyCards")
             QDialog.accept(diag)
             self.checkpoint(_("Delete Empty"))
             self.col.remCards(cids)
